requirements.txt
# PyTorch — CUDA 12.8 (RTX 5070)
--extra-index-url https://download.pytorch.org/whl/cu128
torch
torchvision
torchaudio

# LLM / Fine-Tuning
transformers
datasets
accelerate
scipy
scikit-learn

# Speech-to-Text
SpeechRecognition
azure-cognitiveservices-speech

# Voice Cloning
coqui-tts
sounddevice>=0.4.6
soundfile>=0.12.1
pydub>=0.25.1
av
yt-dlp
edge-tts

# Data Collection
requests
beautifulsoup4
PyPDF2
lxml

# NLP Utilities
nltk
regex
ftfy

# General
numpy
tqdm
python-dotenv